// src/app/usuarios/interfaces/usuario.interfaces.ts
export interface Clientes {
  estado:         string;
  _id:            string;
  nombreCompleto: string;
  identidad:      number;
  numeroTelefono: string;
  fechaIngreso:   string;
  __v:            number;
  correo:         string;
  imagen?:        string;
  idPlan?:        string;
  diasRestantes?: number;
}
// Generated by https://quicktype.io

// export interface Clientes2 {
//   _id:            string;
//   imagen:         string;
//   nombreCompleto: string;
//   identidad:      string;
//   numeroTelefono: string;
//   correo:         string;
//   estado:         Estado;
//   idPlan:         IDPlan;
//   fechaIngreso?:  string;
//   __v:            number;
//   diasRestantes?: number;
// }

// export enum Estado {
//   Activo = "Activo",
//   Inactivo = "Inactivo",
//   Pendiente = "Pendiente",
// }

// export enum IDPlan {
//   The65Dd5E427D9B0513Dce45D5E = "65dd5e427d9b0513dce45d5e",
// }